refactor: rename index component form events to LumaEvents and replace svg icons with Lucide icons
